Improve condition handling in caddy-shim
This commit improves handling of the 'version' variable in the 'caddy-shim.go' file. It removes the 'else' clause from the code and introduces a more specific condition for setting the 'version'. This change ensures that if a version number hasn't been specified, it will default to a predefined number given by 'defaultVersion'. This makes the code cleaner and easier to read.